Breaking out extra_info panel from sidebar.c - stage 1 in refactor, extra globals introduced to sidebar.h to be removed later
