+++ CMakeLists.txt
@@ -452,6 +452,7 @@ set(WIDGET_FILES
     ${PROJECT_SOURCE_DIR}/src/widget/map_editor_tool.c
     ${PROJECT_SOURCE_DIR}/src/widget/minimap.c
     ${PROJECT_SOURCE_DIR}/src/widget/sidebar.c
+    ${PROJECT_SOURCE_DIR}/src/widget/extra_info.c    
     ${PROJECT_SOURCE_DIR}/src/widget/sidebar_editor.c
     ${PROJECT_SOURCE_DIR}/src/widget/top_menu.c
     ${PROJECT_SOURCE_DIR}/src/widget/top_menu_editor.c
