+++ src/widget/sidebar.c
@@ -1,4 +1,5 @@
 #include "sidebar.h"
+#include "extra_info.h"
 
 #include "building/menu.h"
 #include "city/labor.h"
@@ -40,20 +41,8 @@
 #include "window/mission_briefing.h"
 #include "window/overlay_menu.h"
 
-#define SIDEBAR_COLLAPSED_WIDTH 42
-#define SIDEBAR_EXPANDED_WIDTH 162
-#define FILLER_Y_OFFSET 474
-
-#define EXTRA_INFO_LINE_SPACE 32
-#define EXTRA_INFO_HEIGHT_GAME_SPEED 64
-#define EXTRA_INFO_HEIGHT_UNEMPLOYMENT 112
-#define EXTRA_INFO_HEIGHT_RATINGS 272
-#define EXTRA_INFO_TOP_PADDING 10;
-#define EXTRA_INFO_SPEED_PADDING 26;
-#define EXTRA_INFO_UNEMPLOYMENT_TOP_PADDING 20;
-#define EXTRA_INFO_UNEMPLOYMENT_BOTTOM_PADDING 22;
-
 #define SIDEBAR_SLIDE_STEPS 94
+
 
 // sliding sidebar progress to x offset translation
 static const int PROGRESS_TO_X_OFFSET[SIDEBAR_SLIDE_STEPS] = {
@@ -81,7 +70,6 @@ static void button_empire(int param1, int param2);
 static void button_mission_briefing(int param1, int param2);
 static void button_rotate_north(int param1, int param2);
 static void button_rotate(int clockwise, int param2);
-static void button_game_speed(int is_down, int param2);
 
 static image_button buttons_overlays_collapse_sidebar[] = {
     {127, 5, 31, 20, IB_NORMAL, 90, 0, button_collapse_expand, button_none, 0, 0, 1},
@@ -134,32 +122,9 @@ static image_button buttons_top_expanded[] = {
     {123, 184, 33, 22, IB_NORMAL, GROUP_SIDEBAR_BRIEFING_ROTATE_BUTTONS, 9, button_rotate, button_none, 1, 0, 1},
 };
 
-static arrow_button arrow_buttons_speed[] = {
-    {11, 30, 17, 24, button_game_speed, 1, 0},
-    {35, 30, 15, 24, button_game_speed, 0, 0},
-};
 
-typedef struct {
-    int value;
-    int target;
-} objective;
 
-static struct {
-    time_millis slide_start;
-    int progress;
-    int focus_button_for_tooltip;
-    struct {
-        int height;
-        int game_speed;
-        int unemployment_percentage;
-        int unemployment_amount;
-        objective culture;
-        objective prosperity;
-        objective peace;
-        objective favor;
-        objective population;
-    } extra_info;
-} data;
+
 
 static int get_x_offset_expanded(void)
 {
@@ -229,180 +194,6 @@ static void draw_sidebar_filler(int x_offset, int y_offset, int is_collapsed)
     }
 }
 
-static int calculate_extra_info_height(int is_collapsed)
-{
-    if (is_collapsed || !config_get(CONFIG_UI_SIDEBAR_INFO)) {
-        data.extra_info.height = 0;
-    } else {
-        int available_height = screen_height() - FILLER_Y_OFFSET;
-        if (available_height >= EXTRA_INFO_HEIGHT_RATINGS) {
-            data.extra_info.height = EXTRA_INFO_HEIGHT_RATINGS;
-        } else if (available_height >= EXTRA_INFO_HEIGHT_UNEMPLOYMENT) {
-            data.extra_info.height = EXTRA_INFO_HEIGHT_UNEMPLOYMENT;
-        } else if (available_height >= EXTRA_INFO_HEIGHT_GAME_SPEED) {
-            data.extra_info.height = EXTRA_INFO_HEIGHT_GAME_SPEED;
-        } else {
-            data.extra_info.height = 0;
-        }
-    }
-    return data.extra_info.height;
-}
-
-static void set_extra_info_objectives(void)
-{
-    data.extra_info.culture.target = 0;
-    data.extra_info.prosperity.target = 0;
-    data.extra_info.peace.target = 0;
-    data.extra_info.favor.target = 0;
-    data.extra_info.population.target = 0;
-
-    if (scenario_is_open_play()) {
-        return;
-    }
-    if (scenario_criteria_culture_enabled()) {
-        data.extra_info.culture.target = scenario_criteria_culture();
-    }
-    if (scenario_criteria_prosperity_enabled()) {
-        data.extra_info.prosperity.target = scenario_criteria_prosperity();
-    }
-    if (scenario_criteria_peace_enabled()) {
-        data.extra_info.peace.target = scenario_criteria_peace();
-    }
-    if (scenario_criteria_favor_enabled()) {
-        data.extra_info.favor.target = scenario_criteria_favor();
-    }
-    if (scenario_criteria_population_enabled()) {
-        data.extra_info.population.target = scenario_criteria_population();
-    }
-}
-
-static int update_extra_info_value(int value, int *field)
-{
-    if (value == *field) {
-        return 0;
-    } else {
-        *field = value;
-        return 1;
-    }
-}
-
-static int update_extra_info(int height, int is_background)
-{
-    int changed = 0;
-    if (height >= EXTRA_INFO_HEIGHT_GAME_SPEED) {
-        changed |= update_extra_info_value(setting_game_speed(), &data.extra_info.game_speed);
-    }
-    if (height >= EXTRA_INFO_HEIGHT_UNEMPLOYMENT) {
-        changed |= update_extra_info_value(city_labor_unemployment_percentage(), &data.extra_info.unemployment_percentage);
-        changed |= update_extra_info_value(
-            city_labor_workers_unemployed() - city_labor_workers_needed(),
-            &data.extra_info.unemployment_amount
-        );
-    }
-    if (height >= EXTRA_INFO_HEIGHT_RATINGS) {
-        if (is_background) {
-            set_extra_info_objectives();
-        }
-        changed |= update_extra_info_value(city_rating_culture(), &data.extra_info.culture.value);
-        changed |= update_extra_info_value(city_rating_prosperity(), &data.extra_info.prosperity.value);
-        changed |= update_extra_info_value(city_rating_peace(), &data.extra_info.peace.value);
-        changed |= update_extra_info_value(city_rating_favor(), &data.extra_info.favor.value);
-        changed |= update_extra_info_value(city_population(), &data.extra_info.population.value);
-    }
-    return changed;
-}
-
-static void draw_extra_info_objective(int x_offset, int y_offset, int text_group, int text_id, objective *obj, int cut_off_at_parenthesis)
-{
-    if (cut_off_at_parenthesis) {
-        // Exception for Chinese: the string for "population" includes the hotkey " (6)"
-        // To fix that: cut the string off at the '('
-        uint8_t tmp[100];
-        string_copy(lang_get_string(text_group, text_id), tmp, 100);
-        for (int i = 0; i < 100 && tmp[i]; i++) {
-            if (tmp[i] == '(') {
-                tmp[i] = 0;
-                break;
-            }
-        }
-        text_draw(tmp, x_offset + 11, y_offset, FONT_NORMAL_WHITE, 0);
-    } else {
-        lang_text_draw(text_group, text_id, x_offset + 11, y_offset, FONT_NORMAL_WHITE);
-    }
-    font_t font = obj->value >= obj->target ? FONT_NORMAL_GREEN : FONT_NORMAL_RED;
-    int width = text_draw_number(obj->value, '@', "", x_offset + 11, y_offset + 16, font);
-    text_draw_number(obj->target, '(', ")", x_offset + 11 + width, y_offset + 16, font);
-}
-
-
-static void draw_extra_info_panel(int x_offset, int extra_info_height)
-{
-    int y_offset = FILLER_Y_OFFSET;
-
-    // Borders of the extra_info panel - extends to bottom of screen
-    int panel_blocks = extra_info_height / 16;
-    graphics_draw_vertical_line(x_offset, y_offset, y_offset + extra_info_height, COLOR_WHITE);
-    graphics_draw_vertical_line(x_offset + SIDEBAR_EXPANDED_WIDTH - 1, y_offset, y_offset + extra_info_height, COLOR_SIDEBAR);
-    inner_panel_draw(x_offset + 1, y_offset, SIDEBAR_EXPANDED_WIDTH / 16, panel_blocks);
-
-    int y_current_line = y_offset + EXTRA_INFO_TOP_PADDING;
-
-    // Game speed modifier in extra_info panel
-    lang_text_draw(45, 2, x_offset + 11, y_current_line, FONT_NORMAL_WHITE);
-    y_current_line += EXTRA_INFO_SPEED_PADDING;
-
-    text_draw_percentage(data.extra_info.game_speed, x_offset + 60, y_current_line, FONT_NORMAL_GREEN);
-    arrow_buttons_draw(x_offset, y_offset, arrow_buttons_speed, 2);
-    y_current_line += EXTRA_INFO_LINE_SPACE; 
-    
-    // Unemployment info in extra_info panel
-    if (extra_info_height >= EXTRA_INFO_HEIGHT_UNEMPLOYMENT) {
-        lang_text_draw(68, 148, x_offset + 11, y_current_line, FONT_NORMAL_WHITE);
-        y_current_line += EXTRA_INFO_UNEMPLOYMENT_TOP_PADDING;
-
-        int width = text_draw_percentage(data.extra_info.unemployment_percentage, x_offset + 11, y_current_line, FONT_NORMAL_GREEN);
-        text_draw_number(data.extra_info.unemployment_amount, '(', ")", x_offset + 11 + width, y_current_line, FONT_NORMAL_GREEN);
-        y_current_line += EXTRA_INFO_UNEMPLOYMENT_BOTTOM_PADDING;
-    }
-
-    // Objective value info on extra_info panel (culture, prosperity, peace, etc)
-    if (extra_info_height >= EXTRA_INFO_HEIGHT_RATINGS) {
-        draw_extra_info_objective(x_offset, y_current_line, 53, 1, &data.extra_info.culture, 0);
-        y_current_line += EXTRA_INFO_LINE_SPACE;
-
-        draw_extra_info_objective(x_offset, y_current_line, 53, 2, &data.extra_info.prosperity, 0);
-        y_current_line += EXTRA_INFO_LINE_SPACE;
-
-        draw_extra_info_objective(x_offset, y_current_line, 53, 3, &data.extra_info.peace, 0);
-        y_current_line += EXTRA_INFO_LINE_SPACE;
-
-        draw_extra_info_objective(x_offset, y_current_line, 53, 4, &data.extra_info.favor, 0);
-        y_current_line += EXTRA_INFO_LINE_SPACE;
-
-        draw_extra_info_objective(x_offset, y_current_line, 4, 6, &data.extra_info.population, 1);
-        y_current_line += EXTRA_INFO_LINE_SPACE;
-    }
-}
-
-static void draw_extra_info_buttons(int x_offset, int is_collapsed)
-{
-    int extra_info_height = data.extra_info.height;
-    if (!extra_info_height) {
-        return;
-    }
-
-    graphics_set_clip_rectangle(x_offset, TOP_MENU_HEIGHT,
-            screen_width() - x_offset,
-            screen_height() - TOP_MENU_HEIGHT);
-
-    if (update_extra_info(extra_info_height, 0)) {
-        draw_extra_info_panel(x_offset, extra_info_height);
-    } else {
-        arrow_buttons_draw(x_offset, FILLER_Y_OFFSET, arrow_buttons_speed, 2);
-    }
-
-    graphics_reset_clip_rectangle();
-}
 
 static void draw_sidebar_remainder(int x_offset, int is_collapsed)
 {
@@ -498,16 +289,16 @@ int widget_sidebar_handle_mouse(const mouse *m)
     }
     int click = 0;
     int button_id;
-    data.focus_button_for_tooltip = 0;
+    sidebar_data_vals.focus_button_for_tooltip = 0;
     if (city_view_is_sidebar_collapsed()) {
         int x_offset = get_x_offset_collapsed();
         click |= image_buttons_handle_mouse(m, x_offset, 24, button_expand_sidebar, 1, &button_id);
         if (button_id) {
-            data.focus_button_for_tooltip = 12;
+            sidebar_data_vals.focus_button_for_tooltip = 12;
         }
         click |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_collapsed, 12, &button_id);
         if (button_id) {
-            data.focus_button_for_tooltip = button_id + 19;
+            sidebar_data_vals.focus_button_for_tooltip = button_id + 19;
         }
     } else {
         if (widget_minimap_handle_mouse(m)) {
@@ -516,17 +307,17 @@ int widget_sidebar_handle_mouse(const mouse *m)
         int x_offset = get_x_offset_expanded();
         click |= image_buttons_handle_mouse(m, x_offset, 24, buttons_overlays_collapse_sidebar, 2, &button_id);
         if (button_id) {
-            data.focus_button_for_tooltip = button_id + 9;
+            sidebar_data_vals.focus_button_for_tooltip = button_id + 9;
         }
         click |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_expanded, 15, &button_id);
         if (button_id) {
-            data.focus_button_for_tooltip = button_id + 19;
+            sidebar_data_vals.focus_button_for_tooltip = button_id + 19;
         }
         click |= image_buttons_handle_mouse(m, x_offset, 24, buttons_top_expanded, 6, &button_id);
         if (button_id) {
-            data.focus_button_for_tooltip = button_id + 39;
+            sidebar_data_vals.focus_button_for_tooltip = button_id + 39;
         }
-        if (data.extra_info.height >= EXTRA_INFO_HEIGHT_GAME_SPEED) {
+        if (extra_info_height_game_speed_check(sidebar_data_vals.extra_info_vals.height)) {
             click |= arrow_buttons_handle_mouse(m, x_offset, FILLER_Y_OFFSET, arrow_buttons_speed, 2);
         }
     }
@@ -544,7 +335,7 @@ int widget_sidebar_handle_mouse_build_menu(const mouse *m)
 
 int widget_sidebar_get_tooltip_text(void)
 {
-    return data.focus_button_for_tooltip;
+    return sidebar_data_vals.focus_button_for_tooltip;
 }
 
 static void button_overlay(int param1, int param2)
@@ -622,7 +413,7 @@ static void button_rotate(int clockwise, int param2)
     window_invalidate();
 }
 
-static void button_game_speed(int is_down, int param2)
+void button_game_speed(int is_down, int param2)
 {
     if (is_down) {
         setting_decrease_game_speed();
@@ -634,15 +425,15 @@ static void button_game_speed(int is_down, int param2)
 static void update_progress(void)
 {
     time_millis now = time_get_millis();
-    time_millis diff = now - data.slide_start;
-    data.progress = diff / 5;
+    time_millis diff = now - sidebar_data_vals.slide_start;
+    sidebar_data_vals.progress = diff / 5;
 }
 
 static void draw_sliding_foreground(void)
 {
     window_request_refresh();
     update_progress();
-    if (data.progress >= SIDEBAR_SLIDE_STEPS) {
+    if (sidebar_data_vals.progress >= SIDEBAR_SLIDE_STEPS) {
         city_view_toggle_sidebar();
         window_city_show();
         window_draw(1);
@@ -661,9 +452,9 @@ static void draw_sliding_foreground(void)
 
     // draw expanded sidebar on top of it
     if (city_view_is_sidebar_collapsed()) {
-        x_offset_expanded += PROGRESS_TO_X_OFFSET[SIDEBAR_SLIDE_STEPS - data.progress];
+        x_offset_expanded += PROGRESS_TO_X_OFFSET[SIDEBAR_SLIDE_STEPS - sidebar_data_vals.progress];
     } else {
-        x_offset_expanded += PROGRESS_TO_X_OFFSET[data.progress];
+        x_offset_expanded += PROGRESS_TO_X_OFFSET[sidebar_data_vals.progress];
     }
     image_draw(image_base + 1, x_offset_expanded, 24);
     image_buttons_draw(x_offset_expanded, 24, buttons_overlays_collapse_sidebar, 2);
@@ -686,8 +477,8 @@ static void draw_sliding_foreground(void)
 
 static void slide_sidebar(void)
 {
-    data.progress = 0;
-    data.slide_start = time_get_millis();
+    sidebar_data_vals.progress = 0;
+    sidebar_data_vals.slide_start = time_get_millis();
     city_view_start_sidebar_toggle();
     sound_effect_play(SOUND_EFFECT_SIDEBAR);
